2004-06-26  Martin Baulig  <martin@ximian.com>
	* arch/Bfd.cs: Use correct start/end addresses.

	* arch/DwarfReader.cs (DwarfException, DwarfBinaryReader): Don't
	nest these two classes in DwarfReader.

	* arch/DwarfFrameReader.cs: New file.

	* classes/Module.cs (Module.UnwindStack): New abstract method.

svn path=/trunk/debugger/; revision=30439

+++ backends/native/NativePointerObject.cs
@@ -78,7 +78,7 @@ namespace Mono.Debugger.Languages.Native
 
 				long address;
 				if (type.Size == 4)
-					address = BitConverter.ToInt32 (data, 0);
+					address = (uint) BitConverter.ToInt32 (data, 0);
 				else
 					address = BitConverter.ToInt64 (data, 0);
